Use react query for portfolio history
